fix bug

+++ app/expenses/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { group, log } from "console";
-import { useState, useEffect } from "react";
+import { useState, useEffect,useMemo} from "react";
 
 type Transaction = {
   id: string;
@@ -28,23 +28,25 @@ export default function ExpensesPage() {
     useState<Transaction | null>(null);
 
   // calculate Transaction
-  let total = 0;
-  let totalIncome = 0;
-  let totalExpense = 0;
+  // let total = 0;
+  // let totalIncome = 0;
+  // let totalExpense = 0;
 
-  const calculateBalance = () => {
-    transactions.map((e) => {
-      if (e.type === "INCOME") {
-        totalIncome += e.amount;
-      } else if (e.type === "EXPENSE") {
-        totalExpense += e.amount;
-      }
+  // const calculateBalance = () => {
+  //   transactions.map((e) => {
+  //     if (e.type === "INCOME") {
+  //       totalIncome += e.amount;
+  //     } else if (e.type === "EXPENSE") {
+  //       totalExpense += e.amount;
+  //     }
 
-      total = e.type === "INCOME" ? total + e.amount : total - e.amount;
-    });
-  };
+  //     total = e.type === "INCOME" ? total + e.amount : total - e.amount;
+  //   });
+  // };
 
-  calculateBalance();
+  // calculateBalance();
+
+
 
   console.log("Date : ", date);
 
@@ -67,6 +69,10 @@ export default function ExpensesPage() {
     if (type === "-") {
       alert("Enter INCOME or EXPENSE");
       return;
+    }
+    if(amount <= 0){
+      alert("กรอกตัวเลขที่มากกว่า 0 ")
+      return 
     }
 
     const newTransaction = {
@@ -101,6 +107,7 @@ export default function ExpensesPage() {
   };
 
   const deleteTransaction = async (id: string) => {
+    if(!window.confirm("ต้องการลบรายการนี้หรือไม่")) return
     try {
       const res = await fetch(`http://localhost:3001/expenses/${id}`, {
         method: "DELETE",
@@ -117,6 +124,10 @@ export default function ExpensesPage() {
 
   const updateTransaction = async () => {
     if (!editingTransaction) return;
+    if(editingTransaction.amount <= 0){
+      alert("โปรดกรอกตัวเลขมากกว่า 0")
+       return
+    }
 
     try {
       const res = await fetch(
@@ -180,13 +191,42 @@ export default function ExpensesPage() {
     (a, b) => new Date(b).getTime() - new Date(a).getTime(),
   );
 
+
+  const clearFilter = () =>{
+    setSearchTerm("");
+    setFilterStatus("ALL")
+    setDateSearch("");
+  }
+
+  
+  const {totalIncome,totalExpense,totalBalance} = useMemo(()=>{
+
+    return filterTransaction.reduce((acc,curr)=>{
+      if(curr.type === "INCOME"){
+        acc.totalIncome += curr.amount
+        // acc.totalBalance += curr.amount
+      }
+      else if(curr.type === "EXPENSE"){
+        acc.totalExpense += curr.amount
+       
+      }
+      acc.totalBalance = acc.totalIncome - acc.totalExpense
+      return acc
+    },{totalIncome : 0,totalExpense: 0 ,totalBalance : 0} as {
+      totalIncome: number;
+      totalExpense: number;
+      totalBalance: number;
+    }
+  );
+  },[filterTransaction])
+
   console.log("Date ttttt",new Date("2026-03-21").getTime())
   console.log("Group ", groupedTransactions);
   console.log("Group Sort", sortedDates);
   console.log("type", type);
   console.log("Editing Transaction : ", editingTransaction);
   return (
-    <div className="bg-gray-400 w-full h-full min-h-screen flex flex-col items-center ">
+    <div className="bg-gray-400 text-white w-full h-full min-h-screen flex flex-col items-center ">
       <div className="flex flex-col mt-5">
         <h1 className="font-bold text-5xl p-3">Expenses Tracking</h1>
         <div className="bg-gray-700 p-5 rounded-2xl">
@@ -205,8 +245,8 @@ export default function ExpensesPage() {
               <input
                 type="number"
                 className="border border-black-2 rounded-sm p-2 outline-0"
-                value={amount}
-                onChange={(e) => setAmount(Number(e.target.value))}
+                value={amount === 0 ? "" : amount}
+                onChange={(e) => setAmount(e.target.value === "" ? 0 :Number((e.target.value)))}
               />
             </div>
           </div>
@@ -268,25 +308,25 @@ export default function ExpensesPage() {
           <h2>Balance</h2>
           <h3
             className={`p-3 rounded-sm
-            ${total > 0 ? `bg-green-500` : `bg-red-500`}`}
+            ${totalBalance > 0 ? `bg-green-500` : `bg-red-500`}`}
           >
-            {total}
+            {totalBalance.toLocaleString()}
           </h3>
         </div>
         <div className="flex justify-between items-center">
           <div>
             <h2>INCOME</h2>
-            <p>{totalIncome}</p>
+            <p>{totalIncome.toLocaleString()}</p>
           </div>
           <div>
             <h2>EXPENSE</h2>
-            <p>{totalExpense}</p>
+            <p>{totalExpense.toLocaleString()}</p>
           </div>
         </div>
       </div>
       <div className="mt-5 bg-gray-600 p-5 rounded-2xl w-full max-w-md">
         <h1 className="font-bold text-2xl text-center">Transaction</h1>
-        <div className="mt-3 flex flex-col justify-center ">
+        <div className="mt-3 flex flex-col justify-center">
           <input
             type="text"
             className="p-2 border border-white w-full rounded-sm outline-0"
@@ -313,169 +353,18 @@ export default function ExpensesPage() {
               );
             })}
           </div>
-          <input type="date"  value={dateSearch} onChange={(e)=> setDateSearch(e.target.value)} className="p-2 border border-white-2 rounded-xl outline-0 mt-5"/>
+          <div className="flex justify-between mt-5">
+            
+          <input type="date"  value={dateSearch} onChange={(e)=> setDateSearch(e.target.value)} className="p-2 border border-white-2 rounded-xl outline-0 "/>
+            <button onClick={()=> clearFilter()} className="bg-red-400 border border-1 p-2 rounded-sm">clear</button>
+          </div>
         </div>
 
-        {/* <div className="flex flex-col">
-          {filterTransaction.map((e, i) => (
-            <div
-              key={e.id}
-              className={`bg-gray-500 mt-3 p-3 rounded-xl border-l-10
-                ${e.type === "INCOME" ? `border-green-500` : `border-red-500 `}
-
-            `}
-            >
-              <div className="flex justify-between items-center">
-                <div>
-                  <h2 className="text-2xl font-semibold">{e.title}</h2>
-                  <h2>{e.category}</h2>
-                </div>
-
-                <div>
-                  <h2 className="font-semibold text-xl">{e.amount} baht</h2>
-                </div>
-              </div>
-              <div className="flex justify-between">
-                <h2>{e.type}</h2>
-                <h2>{e.date}</h2>
-              </div>
-              <div className="flex justify-end space-x-3.5">
-                <button
-                  className="bg-yellow-500 p-2 px-4 rounded-sm"
-                  onClick={() => {
-                    setEditingTransaction(e);
-                    setToggleEdit(true);
-                  }}
-                >
-                  Edit
-                </button>
-                <button
-                  className="bg-red-500 p-2 px-4 rounded-sm"
-                  onClick={() => deleteTransaction(e.id)}
-                >
-                  Delete
-                </button>
-              </div>
-            </div>
-          ))}
-          {toggleEdit && editingTransaction && (
-            <div className="fixed inset-0 z-50 flex items-center justify-center backdrop-blur-2xl ">
-              <div className="flex flex-col bg-gray-600 p-5 rounded-2xl">
-                <h1 className="text-center text-3xl font-bold">
-                  Editing Transaction
-                </h1>
-
-                <div className="flex space-x-5 mt-5">
-                  <div className="flex flex-col">
-                    <label htmlFor="" className="text-xl font-semibold">
-                      Title
-                    </label>
-                    <input
-                      type="text"
-                      value={editingTransaction.title}
-                      className="border border-black-2 rounded-sm p-2 outline-0"
-                      onChange={(e) =>
-                        setEditingTransaction({
-                          ...editingTransaction,
-                          title: e.target.value,
-                        })
-                      }
-                    />
-                  </div>
-
-                  <div className="flex flex-col">
-                    <label htmlFor="" className="text-xl font-semibold">
-                      Amount
-                    </label>
-                    <input
-                      type="text"
-                      value={editingTransaction.amount}
-                      className="border border-black-2 rounded-sm p-2 outline-0"
-                      onChange={(e) =>
-                        setEditingTransaction({
-                          ...editingTransaction,
-                          amount: Number(e.target.value),
-                        })
-                      }
-                    />
-                  </div>
-                </div>
-                <div className="flex space-x-2  mt-5 jutifly-center items-center">
-                  <div className="flex flex-col">
-                    <label htmlFor="" className="text-xl font-semibold">
-                      Category
-                    </label>
-                    <input
-                      type="text"
-                      value={editingTransaction.category}
-                      className="border border-black-2 rounded-sm p-2 outline-0"
-                      onChange={(e) =>
-                        setEditingTransaction({
-                          ...editingTransaction,
-                          category: e.target.value,
-                        })
-                      }
-                    />
-                  </div>
-
-                  <div className="flex flex-col mx-3">
-                    <label htmlFor="" className="text-xl font-semibold">
-                      Type
-                    </label>
-                    <select
-                      name=""
-                      id=""
-                      value={editingTransaction.type}
-                      className="bg-gray-700 border border-black-2 rounded-sm p-2 outline-0"
-                      onChange={(e) =>
-                        setEditingTransaction({
-                          ...editingTransaction,
-                          type: e.target.value as any,
-                        })
-                      }
-                    >
-                      <option value="INCOME">INCOME</option>
-                      <option value="EXPENSE">EXPENSE</option>
-                    </select>
-                  </div>
-                </div>
-                <div className="mt-2">
-                  <label htmlFor="" className="text-xl font-semibold">
-                    Date
-                  </label>
-                  <input
-                    type="date"
-                    className="flex border border-white p-2 rounded-sm "
-                    value={editingTransaction.date}
-                    onChange={(e) =>
-                      setEditingTransaction({
-                        ...editingTransaction,
-                        date: e.target.value,
-                      })
-                    }
-                  />
-                </div>
-                <div className="flex justify-end mt-3 space-x-3">
-                  <button
-                    className="bg-gray-400 text-black px-3 py-2 rounded-sm cursor-pointer"
-                    onClick={() => setToggleEdit(false)}
-                  >
-                    cancle
-                  </button>
-                  <button
-                    className="bg-yellow-500 px-3 py-2 rounded-sm cursor-pointer"
-                    onClick={() => updateTransaction()}
-                  >
-                    update
-                  </button>
-                </div>
-              </div>
-            </div>
-          )}
-        </div> */}
-
         <div className="flex flex-col mt-4">
-          {sortedDates.map((date) => (
+          {
+           sortedDates.length === 0 ? (
+            <div className="flex text-center justify-center items-center"><p>ไม่พบรายการ</p></div>
+           ) : sortedDates.map((date) => (
             <div key={date} className="mb-6">
               {/* Date (Divider) */}
               <div className="flex justify-between items-center border-b border-gray-500 pb-1 mb-2">
@@ -504,7 +393,7 @@ export default function ExpensesPage() {
 
                       <div>
                         <h2 className="font-semibold text-xl">
-                          {e.amount} baht
+                          {e.amount.toLocaleString()} baht
                         </h2>
                       </div>
                     </div>
@@ -563,13 +452,13 @@ export default function ExpensesPage() {
                       Amount
                     </label>
                     <input
-                      type="text"
-                      value={editingTransaction.amount}
+                      type="number"
+                      value={editingTransaction.amount === 0 ? "" : editingTransaction.amount }
                       className="border border-black-2 rounded-sm p-2 outline-0"
                       onChange={(e) =>
                         setEditingTransaction({
                           ...editingTransaction,
-                          amount: Number(e.target.value),
+                          amount: e.target.value === "" ? 0 : Number(e.target.value),
                         })
                       }
                     />
@@ -646,7 +535,8 @@ export default function ExpensesPage() {
                 </div>
               </div>
             </div>
-          )}
+          )
+          }
         </div>
       </div>
     </div>
